Only show comment notifications once
Postgres doesn't guarantee that messages from LISTEN/NOTIFY will only be
received once by each consumer. This is very obvious at times, with
multiple comment notifications.

This change passes the comment ID to the client so it can track if it
has seen it before - given the fact that comments are immutable
currently there is no need to show a notification for the same comment
ID twice.