Remove the Mistral Lexer
Adding this was an interesting idea, however, it was a bit too specific
for this project. The code has since been moved into Mistral itself - so
it can be used in the documentation.

We will use a trick with requirements.txt and Heroku to include Mistral
in the deploy, without it ever being part of our code base.

In the future it would be good to let users disable this more easily for
custom deploys.